udpate metrics documentation
